+++ core/amber/src/main/python/core/storage/iceberg/iceberg_catalog_instance.py
@@ -1,7 +1,7 @@
 from pyiceberg.catalog import Catalog
 from typing import Optional
 
-from core.storage.iceberg_utils import create_postgres_catalog
+from core.storage.iceberg.iceberg_utils import create_postgres_catalog
 from core.storage.storage_config import StorageConfig
